requirements_otter.txt
absl-py==1.4.0
accelerate==0.21.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
alabaster==0.7.12
alembic==1.8.1
altair==5.0.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
Babel==2.10.1
backcall==0.2.0
backports.functools-lru-cache==1.6.5
beautifulsoup4==4.12.2
blinker==1.6.2
braceexpand==0.1.7
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
cloudpickle==2.1.0
# cmake==3.27.2
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
databricks-cli==0.17.0
datasets==2.14.5
debugpy==1.8.0
decorator==5.1.1
deepspeed==0.10.3
dill==0.3.7
docker==5.0.3
docker-pycreds==0.4.0
docutils==0.17.1
einops==0.6.1
einops-exts==0.0.4
entrypoints==0.4
etils==0.6.0
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.101.1
ffmpy==0.3.1
filelock==3.8.0
Flask==2.2.2
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.32
googleapis-common-protos==1.56.3
gradio==3.40.1
gradio_client==0.4.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.14.0
hjson==3.1.0
horovod==0.27.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
hydra-core==1.2.0
idna==3.4
ijson==3.2.3
imagesize==1.3.0
importlib-metadata==6.8.0
importlib-resources==5.8.0
inflection==0.5.1
ipykernel==6.25.1
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
latexcodec==2.0.1
linkify-it-py==2.0.2
lit==16.0.6
litellm==0.1.400
llvmlite==0.40.1
Mako==1.2.1
Markdown==3.4.4
markdown-it-py==2.2.0
markdown2==2.4.10
MarkupSafe==2.1.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mlflow==2.5.0
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
multipledispatch==0.6.0
multiprocess==0.70.15
natsort==8.4.0
nest-asyncio==1.5.6
networkx==3.1
ninja==1.11.1
nltk==3.8.1
numba==0.57.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.2.2
open-clip-torch==2.20.0
openai==0.27.10
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
orjson==3.9.5
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathtools==0.1.2
peft==0.5.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.3
promise==2.3
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.1
pybtex==0.24.0
pybtex-docutils==1.0.1
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.0.1
pydub==0.25.1
Pygments==2.16.1
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytorch-gradcam==0.2.1
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
querystring-parser==1.2.4
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rlpython==0.10.3
rpds-py==0.9.2
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.11.2
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.30.0
setproctitle==1.3.2
simplejson==3.17.6
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.4.1
Sphinx==4.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-bibtex==2.4.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.40
sqlparse==0.4.2
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tensorflow-datasets==4.6.0
tensorflow-metadata==1.9.0
termcolor==2.3.0
threadpoolctl==3.2.0
tiktoken==0.4.0
timm==0.9.5
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.32.0
triton==2.0.0
trl==0.7.1
typing_extensions==4.7.1
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.0.4
uvicorn==0.23.2
wandb==0.15.9
wcwidth==0.2.6
webdataset==0.2.48
websocket-client==1.3.3
websockets==11.0.3
Werkzeug==2.2.2
# xformers==0.0.21
xxhash==3.3.0
yacs==0.1.8
yajl==0.3.5
yarl==1.9.2
zipp==3.8.0
